+++ ui/src/modules/engage/graphql/queries.ts
@@ -121,6 +121,7 @@ const users = `
         avatar
         fullName
         position
+        operatorPhone
       }
     }
   }