+++ ui/src/modules/settings/integrations/constants.ts
@@ -8,7 +8,7 @@ export const days = [
   { value: 'thursday', label: 'Thursday' },
   { value: 'friday', label: 'Friday' },
   { value: 'saturday', label: 'Saturday' },
-  { value: 'sunday', label: 'Sunday' },
+  { value: 'sunday', label: 'Sunday' }
 ];
 
 export const hours = [
@@ -60,7 +60,7 @@ export const hours = [
   { value: '10:30 PM', label: '10:30 PM' },
   { value: '11:00 PM', label: '11:00 PM' },
   { value: '11:30 PM', label: '11:30 PM' },
-  { value: '11:59 PM', label: '11:59 PM' },
+  { value: '11:59 PM', label: '11:59 PM' }
 ];
 
 export const INTEGRATION_KINDS = {
@@ -88,7 +88,7 @@ export const INTEGRATION_KINDS = {
     { text: 'Facebook post', value: 'facebook-post' },
     {
       text: 'Facebook messenger',
-      value: 'facebook-messenger',
+      value: 'facebook-messenger'
     },
     { text: 'Gmail', value: 'gmail' },
     { text: 'Nylas gmail', value: 'nylas-gmail' },
@@ -105,22 +105,22 @@ export const INTEGRATION_KINDS = {
     { text: 'Viber', value: 'smooch-viber' },
     { text: 'Line', value: 'smooch-line' },
     { text: 'Twilio', value: 'smooch-twilio' },
-    { text: 'WhatsApp', value: 'whatsapp' },
-  ],
+    { text: 'WhatsApp', value: 'whatsapp' }
+  ]
 };
 
 export const FORM_LOAD_TYPES = {
   SHOUTBOX: 'shoutbox',
   POPUP: 'popup',
   EMBEDDED: 'embedded',
-  ALL_LIST: ['', 'shoutbox', 'popup', 'embedded'],
+  ALL_LIST: ['', 'shoutbox', 'popup', 'embedded']
 };
 
 export const FORM_SUCCESS_ACTIONS = {
   EMAIL: 'email',
   REDIRECT: 'redirect',
   ONPAGE: 'onPage',
-  ALL_LIST: ['', 'email', 'redirect', 'onPage'],
+  ALL_LIST: ['', 'email', 'redirect', 'onPage']
 };
 
 export const MAIL_TOOLBARS_CONFIG = [
@@ -135,10 +135,10 @@ export const MAIL_TOOLBARS_CONFIG = [
       'Underline',
       'NumberedList',
       'BulletedList',
-      'Link',
-    ],
+      'Link'
+    ]
   },
-  { name: 'clear', items: ['RemoveFormat'] },
+  { name: 'clear', items: ['RemoveFormat'] }
 ];
 
 export const INTEGRATIONS = [
@@ -151,7 +151,7 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/facebook.png',
     createModal: 'facebook-post',
     category:
-      'All integrations, For support teams, Marketing automation, Social media',
+      'All integrations, For support teams, Marketing automation, Social media'
   },
   {
     name: 'Facebook messenger',
@@ -163,7 +163,7 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/fb-messenger.png',
     createModal: 'facebook-messenger',
     category:
-      'All integrations, For support teams, Messaging, Social media, Conversation',
+      'All integrations, For support teams, Messaging, Social media, Conversation'
   },
   {
     name: 'Messenger',
@@ -174,7 +174,7 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/messenger.png',
     createUrl: '/settings/integrations/createMessenger',
     category:
-      'All integrations, For support teams, For marketing teams, Marketing automation, Conversation',
+      'All integrations, For support teams, For marketing teams, Marketing automation, Conversation'
   },
   {
     name: 'Gmail',
@@ -187,7 +187,7 @@ export const INTEGRATIONS = [
     createModal: 'gmail',
     createUrl: '/settings/integrations/gmail',
     category:
-      'All integrations, For support teams, Email marketing, Marketing automation, Conversation',
+      'All integrations, For support teams, Email marketing, Marketing automation, Conversation'
   },
   {
     name: 'IMAP by Nylas',
@@ -200,7 +200,7 @@ export const INTEGRATIONS = [
     createModal: 'nylas-imap',
     createUrl: '/settings/integrations/nylas-imap',
     category:
-      'All integrations, For support teams, Marketing automation, Email marketing',
+      'All integrations, For support teams, Marketing automation, Email marketing'
   },
   {
     name: 'Office 365 by Nylas',
@@ -213,7 +213,7 @@ export const INTEGRATIONS = [
     createModal: 'nylas-office365',
     createUrl: 'nylas/oauth2/callback',
     category:
-      'All integrations, For support teams, Marketing automation, Email marketing, Conversation',
+      'All integrations, For support teams, Marketing automation, Email marketing, Conversation'
   },
   {
     name: 'Gmail by Nylas',
@@ -226,7 +226,7 @@ export const INTEGRATIONS = [
     createModal: 'nylas-gmail',
     createUrl: 'nylas/oauth2/callback',
     category:
-      'All integrations, For support teams, Email marketing, Marketing automation, Conversation',
+      'All integrations, For support teams, Email marketing, Marketing automation, Conversation'
   },
   {
     name: 'Microsoft Exchange by Nylas',
@@ -239,7 +239,7 @@ export const INTEGRATIONS = [
     createModal: 'nylas-exchange',
     createUrl: '/settings/integrations/nylas-exchange',
     category:
-      'All integrations, For support teams, Email marketing, Marketing automation, Conversation',
+      'All integrations, For support teams, Email marketing, Marketing automation, Conversation'
   },
   {
     name: 'Pop Ups',
@@ -251,7 +251,7 @@ export const INTEGRATIONS = [
     createModal: 'lead',
     createUrl: '/settings/integrations/lead',
     category:
-      'All integrations, For support teams, For marketing teams, Marketing automation, Surveys and Feedback, Pop ups',
+      'All integrations, For support teams, For marketing teams, Marketing automation, Surveys and Feedback, Pop ups'
   },
   {
     name: 'Outlook by Nylas',
@@ -264,7 +264,7 @@ export const INTEGRATIONS = [
     createModal: 'nylas-outlook',
     createUrl: '/settings/integrations/nylas-outlook',
     category:
-      'All integrations, For support teams, Marketing automation, Email marketing',
+      'All integrations, For support teams, Marketing automation, Email marketing'
   },
   {
     name: 'Yahoo by Nylas',
@@ -277,7 +277,7 @@ export const INTEGRATIONS = [
     createModal: 'nylas-yahoo',
     createUrl: '/settings/integrations/nylas-yahoo',
     category:
-      'All integrations, For support teams, Marketing automation, Email marketing, Conversation',
+      'All integrations, For support teams, Marketing automation, Email marketing, Conversation'
   },
   {
     name: 'Knowledge Base',
@@ -289,7 +289,7 @@ export const INTEGRATIONS = [
     createModal: 'knowledgeBase',
     createUrl: '/settings/integrations/knowledgeBase',
     category:
-      'All integrations, For support teams, For marketing teams, Surveys and Feedback',
+      'All integrations, For support teams, For marketing teams, Surveys and Feedback'
   },
   {
     name: 'Call Pro',
@@ -300,17 +300,7 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/callpro.png',
     createModal: 'callpro',
     category:
-      'All integrations, For support teams, Marketing automation, Phone and video, Conversation',
-  },
-  {
-    name: 'Twitter',
-    description: 'Connect Twitter DMs to your Team Inbox',
-    inMessenger: false,
-    isAvailable: false,
-    kind: 'twitter-dm',
-    logo: '/images/integrations/twitter.png',
-    category:
-      'All integrations, For support teams, Marketing automation, Social media, Messaging, Conversation',
+      'All integrations, For support teams, Marketing automation, Phone and video, Conversation'
   },
   {
     name: 'Chatfuel',
@@ -321,7 +311,7 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/chatfuel.png',
     createModal: 'chatfuel',
     category:
-      'All integrations, For support teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, Marketing automation, Messaging, Conversation'
   },
   {
     name: 'Website',
@@ -332,7 +322,7 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/website.png',
     createModal: 'website',
     createUrl: '/settings/integrations/website',
-    category: 'All integrations, For support teams, Marketing automation',
+    category: 'All integrations, For support teams, Marketing automation'
   },
   {
     name: 'WhatsApp',
@@ -342,7 +332,7 @@ export const INTEGRATIONS = [
     kind: 'whatsapp',
     logo: '/images/integrations/whatsapp.png',
     createModal: 'whatsapp',
-    category: 'All integrations, For support teams, Messaging, Conversation',
+    category: 'All integrations, For support teams, Messaging, Conversation'
   },
   {
     name: 'Telegram',
@@ -353,7 +343,7 @@ export const INTEGRATIONS = [
     kind: 'smooch-telegram',
     logo: '/images/integrations/telegram.png',
     createModal: 'smooch-telegram',
-    category: 'All integrations, For support teams, Messaging, Conversation',
+    category: 'All integrations, For support teams, Messaging, Conversation'
   },
   {
     name: 'Viber',
@@ -364,10 +354,10 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/viber.png',
     createModal: 'smooch-viber',
     category:
-      'All integrations, For support teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, Marketing automation, Messaging, Conversation'
   },
   {
-    name: 'Line by Sunshine Conversations',
+    name: 'Line',
     description: 'See and reply to Line messages in your Team Inbox',
     inMessenger: false,
     isAvailable: true,
@@ -375,19 +365,17 @@ export const INTEGRATIONS = [
     logo: '/images/integrations/line.png',
     createModal: 'smooch-line',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Phone and video, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Phone and video, Conversation'
   },
   {
-    name: 'Twilio SMS by Sunshine Conversations',
-    description:
-      'Connect Twilio API for SMS then send and receive text messages anywhere in the world',
+    name: 'Twitter',
+    description: 'Connect Twitter DMs to your Team Inbox',
     inMessenger: false,
-    isAvailable: true,
-    kind: 'smooch-twilio',
-    logo: '/images/integrations/twilio-ipm.png',
-    createModal: 'smooch-twilio',
+    isAvailable: false,
+    kind: 'twitter-dm',
+    logo: '/images/integrations/twitter.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Conversation',
+      'All integrations, For support teams, Marketing automation, Social media, Messaging, Conversation'
   },
   {
     name: 'Wechat',
@@ -397,7 +385,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/wechat.png',
     category:
-      'All integrations, For support teams, Messaging, Marketing automation, Social media, Conversation',
+      'All integrations, For support teams, Messaging, Marketing automation, Social media, Conversation'
   },
   {
     name: 'Twitter post',
@@ -406,7 +394,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/twitter.png',
     category:
-      'All integrations, For support teams, Marketing automation, Social media',
+      'All integrations, For support teams, Marketing automation, Social media'
   },
   {
     name: 'Instagram',
@@ -415,17 +403,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/instagram.png',
     category:
-      'All integrations, For support teams, Messaging, Marketing automation, Social media, Conversation',
-  },
-  {
-    name: 'Aircall',
-    description:
-      'Connect with Aircall and handle calls directly from your Team Inbox',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/aircall.png',
-    category:
-      'All integrations, For support teams, Marketing automation, Phone and video, Social media, Conversation',
+      'All integrations, For support teams, Messaging, Marketing automation, Social media, Conversation'
   },
   {
     name: 'Amazon Lex',
@@ -434,7 +412,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/amazon-lex.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation'
   },
   {
     name: 'Cisco Jabber',
@@ -444,7 +422,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/cisco-jabber.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Phone and video, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Phone and video, Conversation'
   },
   {
     name: 'Cisco Webex',
@@ -454,16 +432,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/cisco-webex.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Phone and video, Conversation',
-  },
-  {
-    name: 'Dashbot',
-    description: 'Connect to bot analytics platform - Dashbot',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/dashbot.png',
-    category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Phone and video, Conversation'
   },
   {
     name: 'Google Dialog Flow',
@@ -472,7 +441,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/google-dialogflow.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Conversation'
   },
   {
     name: 'Google Hangouts',
@@ -482,7 +451,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/google-hangouts.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Phone and video, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Phone and video, Conversation'
   },
   {
     name: 'IBM Watson',
@@ -491,7 +460,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/ibm-watson.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Natural language processing, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Natural language processing, Conversation'
   },
   {
     name: 'Microsoft Luis',
@@ -501,7 +470,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/microsoft.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Natural language processing, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Natural language processing, Conversation'
   },
   {
     name: 'Microsoft Team',
@@ -511,7 +480,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/ms-teams.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Cloud storage, Messaging, Phone and video, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Cloud storage, Messaging, Phone and video, Conversation'
   },
   {
     name: 'Rasa',
@@ -520,7 +489,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/rasa.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Conversation'
   },
   {
     name: 'SAP Conversational AI',
@@ -530,7 +499,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/sap.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation'
   },
   {
     name: 'Slack',
@@ -540,17 +509,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/slack.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Conversation',
-  },
-  {
-    name: 'Twilio IPM',
-    description:
-      'Connect Twilio IPM account then integrating with third party software in your Team Inbox',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/twilio-ipm.png',
-    category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Conversation'
   },
   {
     name: 'WIT.AI',
@@ -559,7 +518,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/wit.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Natural language processing, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Natural language processing, Conversation'
   },
   {
     name: 'Agatha Answers',
@@ -569,7 +528,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/agatha-answers.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation'
   },
   {
     name: 'Asana',
@@ -578,7 +537,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/asana.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Project management',
+      'All integrations, For support teams, For sales teams, For marketing teams, Project management'
   },
   {
     name: 'Asknicely',
@@ -587,7 +546,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/ask-nicely.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Analytics, Marketing automation, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Analytics, Marketing automation, Conversation'
   },
   {
     name: 'Chatbase',
@@ -596,35 +555,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/chatbase.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Analytics, Marketing automation, Messaging, Phone and video, Conversation',
-  },
-  {
-    name: 'Autopilot',
-    description: 'Use the marketing automation section in your Team Inbox',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/autopilot.png',
-    category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Email marketing, Phone and video, Engage',
-  },
-  {
-    name: 'Base',
-    description: 'Connect Base chat then fun unlimited in your Team Inbox',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/base.png',
-    category:
-      'All integrations, For sales teams, CRM, Marketing automation, Social media, Sales pipeline',
-  },
-  {
-    name: 'Chatbot.mn',
-    description:
-      'Connect Chatbot to your Team Inbox and increase customer engagement',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/chatbot.png',
-    category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Analytics, Marketing automation, Messaging, Phone and video, Conversation'
   },
   {
     name: 'Clearbit',
@@ -633,7 +564,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/clearbit.png',
     category:
-      'All integrations, For sales teams, Marketing automation, Contacts',
+      'All integrations, For sales teams, Marketing automation, Contacts'
   },
   {
     name: 'Glassbox',
@@ -643,7 +574,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/glassbox.png',
     category:
-      'All integrations, For marketing teams, Analytics, Marketing automation, Messaging, Phone and video',
+      'All integrations, For marketing teams, Analytics, Marketing automation, Messaging, Phone and video'
   },
   {
     name: 'Doorbell.io',
@@ -653,7 +584,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/doorbell.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Surveys and Feedback, Conversation',
+      'All integrations, For sales teams, For marketing teams, Surveys and Feedback, Conversation'
   },
   {
     name: 'Dropbox',
@@ -662,7 +593,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/dropbox.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Cloud storage',
+      'All integrations, For support teams, For sales teams, For marketing teams, Cloud storage'
   },
   {
     name: 'Fivetran',
@@ -671,7 +602,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/fivetran.png',
-    category: 'All integrations, For marketing teams, Analytics',
+    category: 'All integrations, For marketing teams, Analytics'
   },
   {
     name: 'Full contact',
@@ -681,7 +612,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/full-contact.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Contacts, Data',
+      'All integrations, For sales teams, For marketing teams, Contacts, Data'
   },
   {
     name: 'Github',
@@ -689,7 +620,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/github.png',
-    category: 'All integrations, Tools',
+    category: 'All integrations, Tools'
   },
   {
     name: 'Google calendar',
@@ -699,7 +630,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/google-calendar.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Scheduling, Sales pipeline',
+      'All integrations, For sales teams, For marketing teams, Scheduling, Sales pipeline'
   },
   {
     name: 'Google Drive',
@@ -708,7 +639,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/google-drive.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Cloud storage',
+      'All integrations, For support teams, For sales teams, For marketing teams, Cloud storage'
   },
   {
     name: 'Hubspot import',
@@ -717,7 +648,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/hubspot.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Contacts, CRM, Data, Tools',
+      'All integrations, For sales teams, For marketing teams, Contacts, CRM, Data, Tools'
   },
   {
     name: 'Jira',
@@ -727,16 +658,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/jira.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Project management',
-  },
-  {
-    name: 'Justcall.io',
-    description: 'Connect to JustCall and join cloud-based phone system',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/justcall.png',
-    category:
-      'All integrations, For support teams, Messaging, Marketing automation, Phone and video, Conversation',
+      'All integrations, For sales teams, For marketing teams, Project management'
   },
   {
     name: 'Keen',
@@ -744,7 +666,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/keen.png',
-    category: 'All integrations, For marketing teams, Analytics',
+    category: 'All integrations, For marketing teams, Analytics'
   },
   {
     name: 'LinkedIn',
@@ -754,7 +676,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/linkedin.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Social media',
+      'All integrations, For support teams, For sales teams, For marketing teams, Messaging, Marketing automation, Social media'
   },
   {
     name: 'Loom',
@@ -764,7 +686,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/loom.png',
     category:
-      'All integrations, For support teams, Screen capture, Conversation',
+      'All integrations, For support teams, Screen capture, Conversation'
   },
   {
     name: 'Mailchimp import',
@@ -773,7 +695,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/mailchimp_logo.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Contacts, CRM, Data, Tools',
+      'All integrations, For sales teams, For marketing teams, Contacts, CRM, Data, Tools'
   },
   {
     name: 'ManDrill',
@@ -782,7 +704,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/mandrill.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Email marketing, Marketing automation, Engage',
+      'All integrations, For sales teams, For marketing teams, Email marketing, Marketing automation, Engage'
   },
   {
     name: 'ManyChat',
@@ -792,7 +714,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/manychat.png',
     category:
-      'All integrations, For support teams, Messaging, Marketing automation, Conversation',
+      'All integrations, For support teams, Messaging, Marketing automation, Conversation'
   },
   {
     name: 'Marketo Import',
@@ -800,7 +722,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/marketo-logo.png',
-    category: 'All integrations, For sales teams, For marketing teams, Tools,',
+    category: 'All integrations, For sales teams, For marketing teams, Tools,'
   },
   {
     name: 'Meya',
@@ -810,7 +732,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/meya.png',
     category:
-      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation',
+      'All integrations, For support teams, For sales teams, For marketing teams, Marketing automation, Messaging, Conversation'
   },
   {
     name: 'Mixpanel',
@@ -820,7 +742,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/mix-panel.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Analytics, Messaging, Marketing automation, Social media',
+      'All integrations, For sales teams, For marketing teams, Analytics, Messaging, Marketing automation, Social media'
   },
   {
     name: 'Outlook calendar',
@@ -830,7 +752,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/outlook-calendar.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Scheduling, Sales pipeline',
+      'All integrations, For sales teams, For marketing teams, Scheduling, Sales pipeline'
   },
   {
     name: 'Parabola',
@@ -839,7 +761,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/parabola.png',
-    category: 'All integrations, For marketing teams, Project management',
+    category: 'All integrations, For marketing teams, Project management'
   },
   {
     name: 'Piesync',
@@ -849,15 +771,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/piesync.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Contacts, Data',
-  },
-  {
-    name: 'Pipedrive',
-    description: 'Connect to Pipedrive and boost your close rate',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/pipedrive.png',
-    category: 'All integrations, For sales teams, CRM, Sales pipeline',
+      'All integrations, For sales teams, For marketing teams, Contacts, Data'
   },
   {
     name: 'Salesforce',
@@ -866,7 +780,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/wechat.png',
-    category: 'All integrations, For sales teams, CRM, Sales pipeline',
+    category: 'All integrations, For sales teams, CRM, Sales pipeline'
   },
   {
     name: 'Salesforce import',
@@ -875,7 +789,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/salesforce.png',
-    category: 'All integrations, Tools',
+    category: 'All integrations, Tools'
   },
   {
     name: 'Sendgrid',
@@ -885,7 +799,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/sendgrid.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Email marketing, Engage',
+      'All integrations, For sales teams, For marketing teams, Email marketing, Engage'
   },
   {
     name: 'Sendinblue',
@@ -895,7 +809,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/sendinblue.png',
     category:
-      'All integrations, For marketing teams, Email marketing, Marketing automation, Engage',
+      'All integrations, For marketing teams, Email marketing, Marketing automation, Engage'
   },
   {
     name: 'Shipamax',
@@ -905,7 +819,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/shipamax.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Project management, Conversation',
+      'All integrations, For sales teams, For marketing teams, Project management, Conversation'
   },
   {
     name: 'Stitch',
@@ -914,7 +828,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/stitch.png',
-    category: 'All integrations, For marketing teams, Data',
+    category: 'All integrations, For marketing teams, Data'
   },
   {
     name: 'Talkdesk',
@@ -924,16 +838,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/talkdesk.png',
     category:
-      'All integrations, For support teams, Phone and video, Conversation',
-  },
-  {
-    name: 'Trello',
-    description: 'Sign up and start planning, work with any team',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/trello.png',
-    category:
-      'All integrations, For sales teams, For marketing teams, Project management',
+      'All integrations, For support teams, Phone and video, Conversation'
   },
   {
     name: 'Webhook',
@@ -942,7 +847,7 @@ export const INTEGRATIONS = [
     inMessenger: false,
     isAvailable: false,
     logo: '/images/integrations/webhook.png',
-    category: 'All integrations, Tools',
+    category: 'All integrations, Tools'
   },
   {
     name: 'Youtube',
@@ -952,7 +857,7 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/youtube.jpeg',
     category:
-      'All integrations, For sales teams, For marketing teams, Social media',
+      'All integrations, For sales teams, For marketing teams, Social media'
   },
   {
     name: 'Zapier',
@@ -962,25 +867,8 @@ export const INTEGRATIONS = [
     isAvailable: false,
     logo: '/images/integrations/zapier.png',
     category:
-      'All integrations, For sales teams, For marketing teams, Contacts, CRM',
-  },
-  {
-    name: 'Zoho',
-    description: 'Sign up and transform your business',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/zoho.png',
-    category:
-      'All integrations, For sales teams, CRM, Project management, Sales pipeline',
-  },
-  {
-    name: 'Calendly',
-    description: 'Schedule your first meeting from your Team Inbox',
-    inMessenger: false,
-    isAvailable: false,
-    logo: '/images/integrations/calendly.png',
-    category: 'All integrations, For sales teams, Scheduling, Sales pipeline',
-  },
+      'All integrations, For sales teams, For marketing teams, Contacts, CRM'
+  }
 ];
 
 export const INTEGRATION_FILTERS = [
@@ -990,12 +878,12 @@ export const INTEGRATION_FILTERS = [
       'All integrations',
       'For support teams',
       'For sales teams',
-      'For marketing teams',
-    ],
+      'For marketing teams'
+    ]
   },
   {
     name: 'Works with',
-    items: ['Conversation', 'Sales pipeline', 'Contacts', 'Pop ups', 'Engage'],
+    items: ['Conversation', 'Sales pipeline', 'Contacts', 'Pop ups', 'Engage']
   },
   {
     name: 'Categories',
@@ -1013,7 +901,7 @@ export const INTEGRATION_FILTERS = [
       'Social media',
       'Scheduling',
       'Screen capture',
-      'Surveys and Feedback',
-    ],
-  },
+      'Surveys and Feedback'
+    ]
+  }
 ];